requirements.txt
requests
opencv-python
numpy
websocket-client